requirements.txt
Django>=4.2
djangorestframework>=3.14

pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.12

openpyxl>=3.1
xlrd>=2.0

python-dateutil>=2.8
pytz>=2023.3
